+++ lib/Controller/DynamicRegistrationController.php
@@ -35,10 +35,32 @@ use OCP\AppFramework\Http\Response;
 use OCP\IRequest;
 use OCP\IURLGenerator;
 use OCP\AppFramework\Services\IAppConfig;
+use OCA\OIDCIdentityProvider\Db\AccessTokenMapper;
+use OCA\OIDCIdentityProvider\Db\Client;
+use OCA\OIDCIdentityProvider\Db\ClientMapper;
+use OCA\OIDCIdentityProvider\Db\RedirectUri;
+use OCA\OIDCIdentityProvider\Db\RedirectUriMapper;
+use OCA\OIDCIdentityProvider\Db\LogoutRedirectUri;
+use OCA\OIDCIdentityProvider\Db\LogoutRedirectUriMapper;
+use OCA\OIDCIdentityProvider\Db\Group;
+use OCA\OIDCIdentityProvider\Db\GroupMapper;
+use OCP\Security\ISecureRandom;
+use OCP\AppFramework\Http\Attribute\BruteForceProtection;
+use OCP\AppFramework\Http\Attribute\AnonRateLimit;
 use Psr\Log\LoggerInterface;
 
 class DynamicRegistrationController extends ApiController
 {
+	/** @var ClientMapper */
+	private $clientMapper;
+	/** @var ISecureRandom */
+	private $secureRandom;
+	/** @var AccessTokenMapper  */
+	private $accessTokenMapper;
+	/** @var RedirectUriMapper  */
+	private $redirectUriMapper;
+	/** @var LogoutRedirectUriMapper  */
+	private $logoutRedirectUriMapper;
 	/** @var ITimeFactory */
 	private $time;
 	/** @var Throttler */
@@ -50,9 +72,17 @@ class DynamicRegistrationController extends ApiController
 	/** @var LoggerInterface */
 	private $logger;
 
+	public const VALID_CHARS = 'ABCDEFGHIJKLMNOPQRSTUVWXYZabcdefghijklmnopqrstuvwxyz0123456789';
+	public const NAME_PREFIX = 'DCR-';
+
 	public function __construct(
 					string $appName,
 					IRequest $request,
+					ClientMapper $clientMapper,
+					ISecureRandom $secureRandom,
+					AccessTokenMapper $accessTokenMapper,
+					RedirectUriMapper $redirectUriMapper,
+					LogoutRedirectUriMapper $logoutRedirectUriMapper,
 					ITimeFactory $time,
 					Throttler $throttler,
 					IURLGenerator $urlGenerator,
@@ -61,6 +91,11 @@ class DynamicRegistrationController extends ApiController
 					)
 	{
 		parent::__construct($appName, $request);
+		$this->secureRandom = $secureRandom;
+		$this->clientMapper = $clientMapper;
+		$this->accessTokenMapper = $accessTokenMapper;
+		$this->redirectUriMapper = $redirectUriMapper;
+		$this->logoutRedirectUriMapper = $logoutRedirectUriMapper;
 		$this->time = $time;
 		$this->throttler = $throttler;
         $this->urlGenerator = $urlGenerator;
@@ -71,12 +106,118 @@ class DynamicRegistrationController extends ApiController
 	/**
      * @PublicPage
 	 * @NoCSRFRequired
+	 * @BruteForceProtection(action=oidc_dcr)
+	 * @AnonRateThrottle(limit=10, period=60)
 	 *
 	 * @return JSONResponse
 	 */
-	public function registerClient(): JSONResponse
+	#[AnonRateLimit(limit: 10, period: 60)]
+	#[BruteForceProtection(action: 'oidc_dcr')]
+	public function registerClient(
+		array $redirect_uris = null,
+		string $client_name = null,
+		string $id_token_signed_response_alg = 'RS256',
+		array $response_types = ['code'],
+		): JSONResponse
 	{
-		return null;
+		if ($this->appConfig->getAppValue('dynamic_client_registration', 'false') != 'true') {
+			$this->logger->info('Access to register dynamic client, but functionality disabled.');
+			return new JSONResponse([
+				'error' => 'dynamic_registration_not_allowed',
+				'error_description' => 'Dynamic Client Registration is disabled.',
+			], Http::STATUS_BAD_REQUEST);
+		}
+
+		if ($redirect_uris == null) {
+			$this->logger->info('No redirect uris provided during register dynamic client.');
+			return new JSONResponse([
+				'error' => 'no_redirect_uris_provided',
+				'error_description' => 'Dynamic Client Registration requires redirect_uris to be set.',
+			], Http::STATUS_BAD_REQUEST);
+		}
+
+		if (!is_array($redirect_uris)) {
+			$this->logger->info('No redirect uris array delivered.');
+			return new JSONResponse([
+				'error' => 'no_redirect_uris_provided',
+				'error_description' => 'Dynamic Client Registration requires redirect_uris to be set.',
+			], Http::STATUS_BAD_REQUEST);
+		}
+
+		if (empty($redirect_uris)) {
+			$this->logger->info('No redirect uris array delivered.');
+			return new JSONResponse([
+				'error' => 'no_redirect_uris_provided',
+				'error_description' => 'Dynamic Client Registration requires at least one redirect_uris to be set.',
+			], Http::STATUS_BAD_REQUEST);
+		}
+
+		if ($this->clientMapper->getNumDcrClients() > 100) {
+			$this->logger->info('Maximum number of dynamic registered clients exceeded.');
+			return new JSONResponse([
+				'error' => 'max_num_clients_exceeded',
+				'error_description' => 'Maximum number of dynamic registered clients exceeded.',
+			], Http::STATUS_BAD_REQUEST);
+		}
+
+		$client = new Client();
+		$name = self::NAME_PREFIX . $this->getClientIp();
+		if ($client_name != null) {
+			$name = substr($client_name, 1, 64);
+		}
+		$client->setName($name);
+		$client->setDcr(true);
+		$client->setClientIdentifier($this->secureRandom->generate(64, self::VALID_CHARS));
+		$client->setSecret($this->secureRandom->generate(64, self::VALID_CHARS));
+		$client->setType('confidential');
+
+		if ($id_token_signed_response_alg === 'RS256') {
+			$client->setSigningAlg('RS256');
+		} else {
+			$client->setSigningAlg('HS256');
+		}
+		$response_types_arr = array();
+		array_push($response_types_arr, 'code');
+		$grant_types_arr = array();
+		array_push($grant_types_arr, 'authorization_code');
+		if (in_array('code', $response_types)) {
+			$client->setFlowType('code');
+		} elseif (in_array('id_token', $response_types)){
+			$client->setFlowType('code id_token');
+			array_push($response_types_arr, 'id_token');
+			array_push($grant_types_arr, 'implicit');
+		}
+		$client->setIssuedAt($this->time->getTime());
+
+		$client = $this->clientMapper->insert($client);
+
+		foreach ($redirect_uris as $redirectUri) {
+			$redirectUriObj = new RedirectUri();
+			$redirectUriObj->setClientId($client->getId());
+			$redirectUriObj->setRedirectUri($redirectUri);
+			$redirectUriObj = $this->redirectUriMapper->insert($redirectUriObj);
+		}
+
+		$jsonResponse = [
+			'client_name' => $client->getName(),
+			'client_id' => $client->getClientIdentifier(),
+			'client_secret' => $client->getSecret(),
+			'redirect_uris' => $redirect_uris,
+			'token_endpoint_auth_method' => 'client_secret_post', // Force to use client secret post
+			'response_types' => $response_types_arr,
+			'grant_types' => $grant_types_arr,
+			'client_id_issued_at' => $client->getIssuedAt(),
+			'client_secret_expires_at' => $client->getIssuedAt() + $this->appConfig->getAppValue('client_expire_time', '3600')
+		];
+
+		return new JSONResponse($jsonResponse, Http::STATUS_CREATED);
+	}
+
+	private function getClientIp() {
+		return $_SERVER['HTTP_X_FORWARDED_FOR']
+			?? $_SERVER['REMOTE_ADDR']
+			?? $_SERVER['HTTP_CLIENT_IP']
+			?? $this->secureRandom->generate(64, self::VALID_CHARS);
 	}
 
 }